Resolves issue #10
