+++ composer.json
@@ -19,11 +19,16 @@
         "php": ">=5.3.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.*"
+        "phpunit/phpunit": "^4.8.36"
     },
     "autoload": {
-        "psr-0": {
-            "Treffynnon\\At": "lib"
+        "psr-4": {
+            "Treffynnon\\At\\": "lib/Treffynnon/At"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Treffynnon\\At\\Tests\\": "tests/Treffynnon/At"
         }
     }
 }
