+++ lib/Treffynnon/At/JobAddException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Treffynnon\At;
+/**
+ * Triggered when the addition of a job to the queue has failed.
+ */
+class JobAddException extends \Exception {}
