+++ lib/Treffynnon/At/JobNotFoundException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Treffynnon\At;
+/**
+ * Triggered when a job cannot be found in the queue
+ */
+class JobNotFoundException extends \Exception {}
