+++ lib/Treffynnon/At/UndefinedPropertyException.php
@@ -1,0 +1,7 @@
+<?php
+namespace Treffynnon\At;
+/**
+ * Triggered when attempting to access class property that has not yet been
+ * defined.
+ */
+class UndefinedPropertyException extends \Exception {}
